Update svd file and use updated svd2rust
Currently released versions of svd2rust have some codegen issues
with padding (my fault, I'm afraid).  This commit updates the
svd file to a more recent version from the nrf5 sdk and uses
my fork of svd2rust with the `--nightly` flag to emit "more correct"
accessors for some peripherals.  You'll notice some peripherals are
emitted with unions and that a number of additional write accessors
are emitted.

Since this changes the interface, I've bumped the version in manifest.

I've tweaked the update script to turn on the `-e` flag so that we
don't need to use so many `&&`.